+++ components/drawer/bus/bus.js
@@ -5,13 +5,12 @@ import { normalize } from 'react-native-elements';
 import Icon from 'react-native-vector-icons/SimpleLineIcons';
 import { HaksikAPI, DormitoryAPI } from '../../../tool/jedaero';
 import getWeek from '../../../tool/getWeek';
-import BusTb from '../../../jsons/busschedule.json';
-import BusA from '../../../tool/busA';
-import BusB from '../../../tool/busB';
-import BusRoute from '../../../jsons/bus_stop.json';
+
 import { mainScreen } from '../../styles/busStyle';
 import busCardStyle from '../../styles/busCardStyle';
 import DormCard from './card/DormCard';
+import HaksikCard from './card/HaksikCard';
+import Bustime from './card/BusTime';
 
 export default class Bus extends Component {
     constructor(props) {
@@ -82,90 +81,13 @@ export default class Bus extends Component {
             <SmartBlock name="스마트 출첵" />
             <Bustime name="버스 시간" />
                 {/* <Swiper style={mainScreen.swiperStyle} containerStyle={mainScreen.swiperContainerStyle} showsPagination={false}> */}
-                    <FoodBlock name="오늘의 학식" navigation={this.props.navigation} food={this.state.haksik} onRefresh={() => this.getHaksik(true)}/>
+                    <HaksikCard name="오늘의 학식" navigation={this.props.navigation} meal={this.state.haksik} onRefresh={() => this.getHaksik(true)}/>
                     <DormCard name="오늘의 숙사밥" navigation={this.props.navigation} meal={this.state.dormitory} onRefresh={() => this.getDormitory(true)}/>
                 {/* </Swiper> */}
             </ScrollView>
         )
     }
 }
-
-
-class Bustime extends Component {
-    constructor(props) {
-        super(props);
-        this.state= {
-            selectedIndex: 0,
-            //버스시간 알고리즘 리턴 값
-            A: BusA(BusTb.timeTable.A, 0),
-            B: BusB(BusTb.timeTable.B, 0),
-            // bustop: BusTb.routeName.A
-        };
-        //정거장 목록
-        this.data = BusRoute.routeName.A
-       
-    }
- 
-    // componentDidMount = () => this.buscheck();
-
-    buscheck = () => {
-        setInterval( () => {
-            this.setState({
-                A: BusA(BusTb.timeTable.A, this.state.value),
-                B: BusB(BusTb.timeTable.B, this.state.value)})
-            
-        }, 25000)
-    }
-
-    render() {
-        return(
-            <View style={mainScreen.blockView}>
-               <View style={{...mainScreen.blockViewTitle, backgroundColor: '#334955',}}>
-                  <Text style={mainScreen.blockViewTitleText}>
-                        <Text style={{color: '#d6ecfa'}}>{this.data[this.state.selectedIndex].name}</Text> {this.props.name}
-                  </Text>
-                  {/* <Picker 
-                    ref={instance => this.dropDownPicker = instance} 
-                    data={this.data} 
-                    label={'name'} 
-                    value={'value'}
-                    onValueChange={(value, selectedIndex) => {
-                          this.setState({
-                            selectedIndex,
-                            value,
-                            A: BusA(BusTb.timeTable.A, value),
-                            B: BusB(BusTb.timeTable.B, value)
-                         })
-                            this.buscheck()
-                        }
-                    }/> */}
-                  <TouchableOpacity onPress={() => this.dropDownPicker.setModalVisible(true)}>
-                     <Text style={busCardStyle.busStopViewer}>정류장별 보기</Text>
-                  </TouchableOpacity>
-                </View>
-                <View style={{...mainScreen.blockViewContainer, flexDirection: 'row',}}>
-                {/* A버스 시간 안내 */}
-                    <View style={mainScreen.blockViewContainerMain}>
-                        <Text style={mainScreen.blockTitle}>A</Text>
-                        <Text style={mainScreen.busWay}>반시계방향</Text>
-                    </View>
-                    <View style={mainScreen.blockViewContainerSub}>
-                        <Text style={mainScreen.blockText}>{this.state.A}</Text>
-                    </View>
-                {/* B버스 시간 안내 */}     
-                    <View style={mainScreen.blockViewContainerMain}>
-                        <Text style={mainScreen.blockTitle}>B</Text>
-                        <Text style={mainScreen.busWay}>시계방향</Text>
-                    </View>
-                    <View style={mainScreen.blockViewContainerSub}>
-                        <Text style={mainScreen.blockText}>{this.state.B}</Text>
-                    </View>
-                </View>
-            </View>
-        )
-    }
-}
-
 
 class SmartBlock extends Component {
     constructor(props) {
@@ -197,91 +119,3 @@ render() {
     }
 }
 
-class FoodBlock extends Component {
-    constructor(props) {
-        super(props);
-        this.state={};
-    }
-
-
-    static getDerivedStateFromProps = (nextProps, prevState) => {
-        return FoodBlock.buildData(nextProps.food)
-    }
-
-    static buildData(meal) {
-        if(meal) {
-            let currentDate = new Date();
-            currentDate = currentDate.getDay();
-            let food;
-            switch(currentDate) {
-                case 1:
-                    food = meal.mealMon;
-                    break;
-                case 2:
-                    food = meal.mealTue;
-                    break;
-                case 3:
-                    food = meal.mealWed;
-                    break;
-                case 4:
-                    food = meal.mealThu;
-                    break;
-                case 5:
-                    food = meal.mealFri;
-                    break;
-                case 6:
-                    food = meal.mealSat ? meal.mealSat : meal.mealMon;
-                    break;
-                case 0:
-                    food = meal.mealSun ? meal.mealSun : meal.mealMon;
-                default:
-                    food = meal.mealMon;
-                    break;
-            }
-            let combo = food.combo.split("\n");
-            let special = food.special.split("\n");
-            let dinner = food.dinner.split("\n");
-            let western = food.western.split("\n");
-
-            return { food: {
-                combo: combo.join(' '),
-                special: special.join(' '),
-                dinner: dinner.join(' '),
-                western: western.join(' ')
-            }};
-        }
-        return {};
-    }
-
-    componentDidMount = () => {
-        this.setState(FoodBlock.buildData(this.props.food));
-    }
-
-    render() {
-        const week = new Array('HaksikMon', 'HaksikMon', 'HaksikTue', 'HaksikWed', 'HaksikThu', 'HaksikFri', 'HaksikMon');
-        let today = new Date().getDay();
-        let day = week[today];
-
-        const { navigation } = this.props;
-        return (
-            <View style={mainScreen.blockView}>
-                <View style={{...mainScreen.blockViewTitle, backgroundColor: '#334955',}}>
-                    <Text style={mainScreen.blockViewTitleText}>{this.props.name}</Text>
-                    <TouchableOpacity
-                        onPress={this.props.onRefresh}
-                    >
-                        <Icon name="refresh" color="#ffffff" size={normalize(16)} />
-                    </TouchableOpacity>
-                </View>
-                <TouchableOpacity onPress = {() => {navigation.navigate(day)}}>
-                <View style={{...mainScreen.foodViewBlockContainer, flexDirection: 'column',}}>
-                    <Text style={mainScreen.foodBlockContainerTitle}>정식 </Text><Text numberOfLines={1} style={busCardStyle.foodBlockContainerText}>{this.state.food? this.state.food.combo+'\n' : '없어요\n'}</Text>
-                    <Text style={mainScreen.foodBlockContainerTitle}>특식 </Text><Text numberOfLines={1} style={busCardStyle.foodBlockContainerText}>{this.state.food? this.state.food.special+'\n' : '없어요\n'}</Text>
-                    <Text style={mainScreen.foodBlockContainerTitle}>양식 </Text><Text numberOfLines={1} style={busCardStyle.foodBlockContainerText}>{this.state.food? this.state.food.western+'\n' : '없어요\n'}</Text>
-                    <Text style={mainScreen.foodBlockContainerTitle}>저녁 </Text><Text numberOfLines={1} style={busCardStyle.foodBlockContainerText}>{this.state.food? this.state.food.dinner+'\n' : '없어요\n'}</Text>
-                </View>
-                </TouchableOpacity>
-            </View>
-        )
-    }
-}
