+++ components/navigations/MainBottomTabNavigation.js
@@ -4,33 +4,35 @@ import { jedaeroBottomTabNavigationConfig } from './navigationConfigs';
 import Icon from 'react-native-vector-icons/MaterialCommunityIcons';
 import { BusStackNavigation, FoodListStackNavigation, LibrarySearchStackNavigation, ScheduleStackNavigation } from './stackNavigations';
 
+const iconSize = 24
+
 const TabNav = createMaterialBottomTabNavigator({
     Bus: {
         screen: BusStackNavigation,
         navigationOptions: {
-            title: '순환버스',
-            tabBarIcon: ({tintColor}) => (<Icon name="home-outline" color={tintColor} size={30} style={{width: 30, height: 60}}/>)
+            title: '홈',
+            tabBarIcon: ({tintColor}) => (<Icon name="home-outline" color={tintColor} size={iconSize} style={{width: iconSize, height: iconSize}}/>)
         }
     },
     Food: {
         screen: FoodListStackNavigation,
         navigationOptions: {
             title: '뭐먹을까',
-            tabBarIcon: ({tintColor}) => (<Icon name="food-fork-drink" color={tintColor} size={30} style={{width: 30, height: 60}}/>)
+            tabBarIcon: ({tintColor}) => (<Icon name="food-fork-drink" color={tintColor} size={iconSize} style={{width: iconSize, height: iconSize}}/>)
         }
     },
     ScheduleMain: {
         screen: ScheduleStackNavigation,
         navigationOptions: {
             title: '학사일정',
-            tabBarIcon: ({tintColor}) => (<Icon name="clipboard-outline" color={tintColor} size={30} style={{width: 30, height: 60}}/>)
+            tabBarIcon: ({tintColor}) => (<Icon name="clipboard-outline" color={tintColor} size={iconSize} style={{width: iconSize, height: iconSize}}/>)
         }
     },    
     LibrarySearch: {
         screen: LibrarySearchStackNavigation,
         navigationOptions: {
             title: '도서검색',
-            tabBarIcon: ({tintColor}) => (<Icon name="book-open-variant" color={tintColor} size={30} style={{width: 30, height: 60}}/>)
+            tabBarIcon: ({tintColor}) => (<Icon name="book-open-variant" color={tintColor} size={iconSize} style={{width: iconSize, height: iconSize}}/>)
         }
     },
 }, jedaeroBottomTabNavigationConfig)
